requirements.txt
# Core Dependencies
numpy>=1.26.0
torch>=2.0.0
sentence-transformers>=3.0.0
transformers>=4.30.0
faiss-cpu>=1.7.4
llama-cpp-python>=0.3.0

# Document Processing
PyMuPDF>=1.24.0
pillow>=10.0.0
tqdm>=4.66.0

# GUI & Packaging
flet>=0.21.0
PySide6>=6.7.0
pyinstaller>=6.0.0
shiboken6>=6.7.0

# Utils
PyYAML>=6.0.0
requests>=2.31.0
huggingface-hub>=0.19.0
